+++ ambix_rotator_z/Source/PluginProcessor.cpp
@@ -443,7 +443,7 @@ void Ambix_rotator_zAudioProcessor::setStateInformation (const void* data, int s
         if (xmlState->hasTagName ("MYPLUGINSETTINGS"))
         {
             for (int i=0; i < getNumParameters(); i++) {
-                setParameter(i, xmlState->getDoubleAttribute(String(i)));
+                setParameter(i, (float)xmlState->getDoubleAttribute(String(i)));
             }
         }
